# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TasteTracker API",
        "description": "API REST del restaurante para gestionar menú, pedidos, clientes y empleados.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal"
        }
    ],
    "paths": {
        "/api/v1/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar todos los clientes",
                "description": "Obtiene el listado completo de clientes ordenado por nombre.",
                "operationId": "275ddabe0ff3df657edd035747f2c43a",
                "responses": {
                    "200": {
                        "description": "Listado obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Cliente"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear un nuevo cliente",
                "description": "Crea un nuevo registro de cliente.",
                "operationId": "d82cb60437b288da7843df7826fad22f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/v1/clientes/{cliente}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener un cliente",
                "description": "Devuelve los detalles de un cliente específico.",
                "operationId": "ab4ce3009d7508c93f86a31ca7afca31",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar un cliente",
                "description": "Actualiza los datos de un cliente.",
                "operationId": "d73c1716e55d5b6be632db863d3fda4d",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar un cliente",
                "description": "Elimina un registro de cliente.",
                "operationId": "49710823fcda170cb253efe37c6339e6",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar parcialmente un cliente",
                "description": "Actualiza uno o más campos de un cliente.",
                "operationId": "35fb29443616cb1ce79ea894eb5c0c84",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Cliente": {
                "required": [
                    "nombre_cliente"
                ],
                "properties": {
                    "id_cliente": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nombre_cliente": {
                        "type": "string",
                        "example": "Ana López"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "555-1234"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Calle 1 #123, Col. Centro"
                    }
                },
                "type": "object"
            },
            "ClienteCreate": {
                "required": [
                    "nombre_cliente"
                ],
                "properties": {
                    "nombre_cliente": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Ana López"
                    },
                    "telefono": {
                        "type": "string",
                        "maxLength": 30,
                        "example": "555-1234"
                    },
                    "direccion": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Calle 1 #123, Col. Centro"
                    }
                },
                "type": "object"
            },
            "ClienteUpdate": {
                "properties": {
                    "nombre_cliente": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Ana M. López"
                    },
                    "telefono": {
                        "type": "string",
                        "maxLength": 30,
                        "example": "555-9876"
                    },
                    "direccion": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Av. Reforma 456"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}